Allow RO to open before RW
Type raw in the Io module is now of type option raw. When a RO instance
is opened first and there are no files on disks, then its raw field is
set to None. RO will try to open the file only the first time it needs
to read it.

RO instances do not refill their memory caches for Dict and the branch
store when opened, but instead only when the first read occurs. To
ensure this, the offset is set to 0 when a file is opened.